Fix auto loading not picking up prometheus.rb
The Zeitwork mode is failing to load the prometheus.rb file in
production environments. Even though other .rb files are working fine.

As a temporary fix, we'll move Promtheus within the Metrics module
